EDIT mpv fix stutter, merge input w latest
Fix stuttering that started happening with the latest mpv. This seems to
be due to video framerate differing from the monitor framerate. I guess
it used to default to resampling to fix this, but now you have to
explicitly define it.

Merged in the latest input.conf from mpv master. Also added alt-s hotkey
for shuffling the playlist.